requirements.txt
numpy
open3d
scipy
tqdm
matplotlib
torch
torchvision
tensorboard~=2.8.0
tabulate
seaborn
networkx
protobuf==3.20.*
trimesh
prettytable
absl-py==1.3.0
appdirs==1.4.4
APScheduler==3.6.3
asttokens==2.0.8
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
black==22.10.0
Brotli==1.0.9
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cmake==3.24.1.1
colorama==0.4.5
ConfigArgParse==1.5.3
contourpy==1.0.5
cycler==0.11.0
dash==2.6.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.1.1
fastjsonschema==2.16.2
flake8==5.0.4
Flask==2.2.2
Flask-Compress==1.13
flatbuffers==22.10.26
fonttools==4.37.4
gast==0.4.0
google-auth==2.13.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.50.0
h5py==3.7.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.16.2
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter-core==4.11.1
jupyter_client==7.4.4
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mypy-extensions==0.4.3
nbformat==5.5.0
nest-asyncio==1.5.6
networkx==2.8.7
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
plotly==5.10.0
prettytable==3.5.0
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-telegram-bot==13.14
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.2
seaborn==0.12.1
six==1.16.0
sklearn==0.0
stack-data==0.5.1
tabulate==0.9.0
telegram-send==0.34
tenacity==8.1.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
tomli==2.0.1
torch==1.12.0+cu116
torch-tb-profiler==0.4.0
torchvision==0.13.0+cu116
tornado==6.1
tqdm==4.64.1
traitlets==5.5.0
trimesh==3.15.5
typing_extensions==4.4.0
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.9.0
